[Gold II] Title: 가장 긴 증가하는 부분 수열 3, Time: 276 ms, Memory: 8924 KB -BaekjoonHub

+++ 백준/Gold/12738. 가장 긴 증가하는 부분 수열 3/가장 긴 증가하는 부분 수열 3.cc
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int n;
-long long int A[1000000];
-long long int R[1000000];
+int A[1000000];
+int R[1000000];
 int main() {
 	scanf("%d", &n);
-	for (int i = 0; i < n; i++) scanf("%lld", &A[i]);
+	for (int i = 0; i < n; i++) scanf("%d", &A[i]);
 	int r = 0;
 
 	for (int i = 0; i < n; i++)
